New private module rand::seq::iterator
